requirements.txt
pyyaml
python-box
pydantic-ai
litellm
pillow
torch
pandas